Added TCP connectivity tests for Cloud SQL

+++ spec/kokoro-run-all.sh
@@ -132,7 +132,7 @@ if [[ $E2E = "true" ]]; then
   mkdir /cloudsql && chmod 0777 /cloudsql
 
   # Start Cloud SQL Proxy.
-  /cloud_sql_proxy -dir=/cloudsql -credential_file=$GOOGLE_APPLICATION_CREDENTIALS &
+  /cloud_sql_proxy -instances=${INSTANCE_CONNECTION_NAME}=tcp:5432,${INSTANCE_CONNECTION_NAME} -dir=/cloudsql -credential_file=$GOOGLE_APPLICATION_CREDENTIALS &
   export CLOUD_SQL_PROXY_PROCESS_ID=$!
   trap "kill $CLOUD_SQL_PROXY_PROCESS_ID || true" EXIT
 fi
@@ -157,7 +157,12 @@ if [[ $RUN_ALL_TESTS = "1" ]]; then
 
     start_time="$(date -u +%s)"
 
-    (bundle update && bundle exec rspec --format documentation --format RspecJunitFormatter --out sponge_log.xml | tee sponge_log.log) || set_failed_status
+    if [[ -f "${REPO_DIRECTORY}/${PRODUCT}/bin/run_tests" ]]; then
+      (bundle update && bin/run_tests) || set_failed_status
+    else
+      (bundle update && bundle exec rspec --format documentation --format RspecJunitFormatter --out sponge_log.xml | tee sponge_log.log) || set_failed_status
+    fi
+
 
     if [[ $E2E = "true" ]]; then
       # Clean up deployed version
